{{cookiecutter.project_name}}/requirements/common.txt
Django==1.8.*
dj-database-url==0.3.*
dj-static==0.0.*
psycopg2==2.6.*
SQLAlchemy==1.0.*
django-postgrespool==0.3.*
django-admin-bootstrapped==2.5.*
python-dotenv==0.1.*
django-webpack-loader==0.1.*
whitenoise==2.0.*
requests==2.7.*
python-social-auth==0.2.*
bcrypt==2.0.*
